Don't depend on the existence of Object class for checking whether exceptions work.
Nicola: this test passes, but then configure decides to override the result because it thinks that my gcc won't work, even after testing that it does.  My autoconf-fu is weak, so maybe you could fix this?  I'm pretty sure the test is wrong anyway, since I was using native exceptions with the GCC runtime that shipped with 4.2.1 (the last GPLv2 version), which is <4.3, before I started working on libobjc2...



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/tools/make/trunk@32043 72102866-910b-0410-8b05-ffd578937521