.git_allowed_signers: Add explicit local verification method for SSH-key-based sigs.
Without this file, the verification for commits depends solely on keys uploaded through hosting platform UIs. The action to declare trusted keys is also made into an explicit action on the repository.

Ideally, this would belong in a separate repository containing trusted maintainer identities in general. This could in the future include hosting-platform-provided identities, auto-generated, but for verification and historical records we should keep this separate.