TestFramework: Add gstest_to_junit_xml.py.
This is currently unused in gnustep-tests, and it is not installed. The
intention is to have a representation of tests run under gnustep-tests
in the JUnit test report XML format. This is a format ingested and
understood by many utilities, so having a way to trigger gnustep-tests
to produce it would be beneficial.

Ideally, the framework itself would produce much more structured output, that
is easier to parse; or it would support creating the .xml natively. The next
best thing would be to process tests.log; however, the simplest approach was
to process tests.sum. This can then trivially be plugged into gnustep-tests,
controlled by an environment variable.

The prototype is done in Python 3, specifically under Python 3.11. Switching
to Objective-C does not necessarily introduce a dependency loop, as it is
perfectly valid to run tests only after GNUstep Base has already been
installed.

Issue: #53
Bug: 53